requirements-full.txt
cat > requirements-full.txt << 'EOF'
streamlit>=1.36.0
openai>=1.40.0
numpy>=1.26.0
pdfplumber>=0.11.0
tqdm>=4.66.0
python-dotenv>=1.0.1
EOF